preston's feedback moving unit tests to appropriate file
