+++ validator/client/BUILD.bazel
@@ -106,6 +106,7 @@ go_test(
     srcs = [
         "aggregate_test.go",
         "attest_test.go",
+        "duties_test.go",
         "health_monitor_test.go",
         "key_reload_test.go",
         "log_test.go",
