+++ validator/client/validator_test.go
@@ -17,7 +17,6 @@ import (
 	"time"
 
 	grpcutil "github.com/OffchainLabs/prysm/v7/api/grpc"
-	"github.com/OffchainLabs/prysm/v7/api/server/structs"
 	"github.com/OffchainLabs/prysm/v7/async/event"
 	"github.com/OffchainLabs/prysm/v7/cmd/validator/flags"
 	"github.com/OffchainLabs/prysm/v7/config/features"
@@ -33,7 +32,6 @@ import (
 	validatorpb "github.com/OffchainLabs/prysm/v7/proto/prysm/v1alpha1/validator-client"
 	"github.com/OffchainLabs/prysm/v7/testing/assert"
 	"github.com/OffchainLabs/prysm/v7/testing/require"
-	"github.com/OffchainLabs/prysm/v7/testing/util"
 	validatormock "github.com/OffchainLabs/prysm/v7/testing/validator-mock"
 	"github.com/OffchainLabs/prysm/v7/validator/accounts/wallet"
 	"github.com/OffchainLabs/prysm/v7/validator/client/iface"
@@ -353,286 +351,6 @@ func TestWaitSync_Syncing(t *testing.T) {
 	).Return(&ethpb.SyncStatus{Syncing: false}, nil)
 
 	require.NoError(t, v.WaitForSync(t.Context()))
-}
-
-func TestUpdateDuties_DoesNothingWhenNotEpochStart_AlreadyExistingAssignments(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	v := validator{
-		km:              newMockKeymanager(t, randKeypair(t)),
-		validatorClient: client,
-		duties: &ethpb.ValidatorDutiesContainer{
-			CurrentEpochDuties: []*ethpb.ValidatorDuty{
-				{
-					AttesterSlot:   10,
-					CommitteeIndex: 20,
-				},
-			},
-			NextEpochDuties: []*ethpb.ValidatorDuty{
-				{
-					AttesterSlot:   10,
-					CommitteeIndex: 20,
-				},
-			},
-		},
-	}
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Times(1)
-
-	assert.NoError(t, v.UpdateDuties(t.Context()), "Could not update assignments")
-}
-
-func TestUpdateDuties_ReturnsError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	v := validator{
-		validatorClient: client,
-		km:              newMockKeymanager(t, randKeypair(t)),
-		duties: &ethpb.ValidatorDutiesContainer{
-			CurrentEpochDuties: []*ethpb.ValidatorDuty{
-				{
-					CommitteeIndex: 1,
-				},
-			},
-		},
-	}
-
-	expected := errors.New("bad")
-
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Return(nil, expected)
-
-	assert.ErrorContains(t, expected.Error(), v.UpdateDuties(t.Context()))
-	assert.Equal(t, (*ethpb.ValidatorDutiesContainer)(nil), v.duties, "Assignments should have been cleared on failure")
-}
-
-func TestUpdateDuties_OK(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	resp := &ethpb.ValidatorDutiesContainer{
-		CurrentEpochDuties: []*ethpb.ValidatorDuty{
-			{
-				AttesterSlot:    params.BeaconConfig().SlotsPerEpoch,
-				ValidatorIndex:  200,
-				CommitteeIndex:  100,
-				CommitteeLength: 4,
-				PublicKey:       []byte("testPubKey_1"),
-				ProposerSlots:   []primitives.Slot{params.BeaconConfig().SlotsPerEpoch + 1},
-			},
-		},
-	}
-	v := validator{
-		km:              newMockKeymanager(t, randKeypair(t)),
-		validatorClient: client,
-	}
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Return(resp, nil)
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-
-	client.EXPECT().SubscribeCommitteeSubnets(
-		gomock.Any(),
-		gomock.Any(),
-		gomock.Any(),
-	).DoAndReturn(func(_ context.Context, _ *ethpb.CommitteeSubnetsSubscribeRequest, _ []*ethpb.ValidatorDuty) (*emptypb.Empty, error) {
-		wg.Done()
-		return nil, nil
-	})
-
-	require.NoError(t, v.UpdateDuties(t.Context()), "Could not update assignments")
-
-	util.WaitTimeout(&wg, 2*time.Second)
-
-	assert.Equal(t, params.BeaconConfig().SlotsPerEpoch+1, v.duties.CurrentEpochDuties[0].ProposerSlots[0], "Unexpected validator assignments")
-	assert.Equal(t, params.BeaconConfig().SlotsPerEpoch, v.duties.CurrentEpochDuties[0].AttesterSlot, "Unexpected validator assignments")
-	assert.Equal(t, resp.CurrentEpochDuties[0].CommitteeIndex, v.duties.CurrentEpochDuties[0].CommitteeIndex, "Unexpected validator assignments")
-	assert.Equal(t, resp.CurrentEpochDuties[0].ValidatorIndex, v.duties.CurrentEpochDuties[0].ValidatorIndex, "Unexpected validator assignments")
-}
-
-func TestUpdateDuties_OK_FilterBlacklistedPublicKeys(t *testing.T) {
-	hook := logTest.NewGlobal()
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	numValidators := 10
-	km := genMockKeymanager(t, numValidators)
-	blacklistedPublicKeys := make(map[[fieldparams.BLSPubkeyLength]byte]bool)
-	for _, k := range km.keys {
-		blacklistedPublicKeys[k] = true
-	}
-	v := validator{
-		km:                 km,
-		validatorClient:    client,
-		blacklistedPubkeys: blacklistedPublicKeys,
-	}
-
-	resp := &ethpb.ValidatorDutiesContainer{
-		CurrentEpochDuties: []*ethpb.ValidatorDuty{},
-	}
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Return(resp, nil)
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-	client.EXPECT().SubscribeCommitteeSubnets(
-		gomock.Any(),
-		gomock.Any(),
-		gomock.Any(),
-	).DoAndReturn(func(_ context.Context, _ *ethpb.CommitteeSubnetsSubscribeRequest, _ []*ethpb.ValidatorDuty) (*emptypb.Empty, error) {
-		wg.Done()
-		return nil, nil
-	})
-
-	require.NoError(t, v.UpdateDuties(t.Context()), "Could not update assignments")
-
-	util.WaitTimeout(&wg, 2*time.Second)
-
-	for range blacklistedPublicKeys {
-		assert.LogsContain(t, hook, "Not including slashable public key")
-	}
-}
-
-func TestUpdateDuties_AllValidatorsExited(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	resp := &ethpb.ValidatorDutiesContainer{
-		CurrentEpochDuties: []*ethpb.ValidatorDuty{
-			{
-				AttesterSlot:    params.BeaconConfig().SlotsPerEpoch,
-				ValidatorIndex:  200,
-				CommitteeIndex:  100,
-				CommitteeLength: 4,
-				PublicKey:       []byte("testPubKey_1"),
-				ProposerSlots:   []primitives.Slot{params.BeaconConfig().SlotsPerEpoch + 1},
-				Status:          ethpb.ValidatorStatus_EXITED,
-			},
-			{
-				AttesterSlot:    params.BeaconConfig().SlotsPerEpoch,
-				ValidatorIndex:  201,
-				CommitteeIndex:  101,
-				CommitteeLength: 4,
-				PublicKey:       []byte("testPubKey_2"),
-				ProposerSlots:   []primitives.Slot{params.BeaconConfig().SlotsPerEpoch + 1},
-				Status:          ethpb.ValidatorStatus_EXITED,
-			},
-		},
-	}
-	v := validator{
-		km:              newMockKeymanager(t, randKeypair(t)),
-		validatorClient: client,
-	}
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Return(resp, nil)
-
-	err := v.UpdateDuties(t.Context())
-	require.ErrorContains(t, ErrValidatorsAllExited.Error(), err)
-
-}
-
-func TestUpdateDuties_Distributed(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	// Start of third epoch.
-	slot := 2 * params.BeaconConfig().SlotsPerEpoch
-	keys := randKeypair(t)
-	resp := &ethpb.ValidatorDutiesContainer{
-		CurrentEpochDuties: []*ethpb.ValidatorDuty{
-			{
-				AttesterSlot:   slot, // First slot in epoch.
-				ValidatorIndex: 200,
-				CommitteeIndex: 100,
-				PublicKey:      keys.pub[:],
-				Status:         ethpb.ValidatorStatus_ACTIVE,
-			},
-		},
-		NextEpochDuties: []*ethpb.ValidatorDuty{
-			{
-				AttesterSlot:   slot + params.BeaconConfig().SlotsPerEpoch, // First slot in next epoch.
-				ValidatorIndex: 200,
-				CommitteeIndex: 100,
-				PublicKey:      keys.pub[:],
-				Status:         ethpb.ValidatorStatus_ACTIVE,
-			},
-		},
-	}
-
-	v := validator{
-		km:              newMockKeymanager(t, keys),
-		validatorClient: client,
-		distributed:     true,
-	}
-
-	sigDomain := make([]byte, 32)
-
-	client.EXPECT().Duties(
-		gomock.Any(),
-		gomock.Any(),
-	).Return(resp, nil)
-
-	client.EXPECT().DomainData(
-		gomock.Any(), // ctx
-		gomock.Any(), // epoch
-	).Return(
-		&ethpb.DomainResponse{SignatureDomain: sigDomain},
-		nil, /*err*/
-	)
-
-	client.EXPECT().AggregatedSelections(
-		gomock.Any(),
-		gomock.Any(), // fill this properly
-	).Return(
-		[]iface.BeaconCommitteeSelection{
-			{
-				SelectionProof: make([]byte, 32),
-				Slot:           slot,
-				ValidatorIndex: 200,
-			},
-			{
-				SelectionProof: make([]byte, 32),
-				Slot:           slot + params.BeaconConfig().SlotsPerEpoch,
-				ValidatorIndex: 200,
-			},
-		},
-		nil,
-	)
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-
-	client.EXPECT().SubscribeCommitteeSubnets(
-		gomock.Any(),
-		gomock.Any(),
-		gomock.Any(),
-	).DoAndReturn(func(_ context.Context, _ *ethpb.CommitteeSubnetsSubscribeRequest, _ []*ethpb.ValidatorDuty) (*emptypb.Empty, error) {
-		wg.Done()
-		return nil, nil
-	})
-
-	require.NoError(t, v.UpdateDuties(t.Context()), "Could not update assignments")
-	util.WaitTimeout(&wg, 2*time.Second)
-	require.Equal(t, 2, len(v.attSelections))
 }
 
 func TestRolesAt_OK(t *testing.T) {
@@ -2859,109 +2577,6 @@ func TestUpdateValidatorStatusCache(t *testing.T) {
 	assert.NoError(t, err)
 	// make sure the value is 0
 	assert.Equal(t, 0, len(v.pubkeyToStatus))
-}
-
-func TestValidator_CheckDependentRoots(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	ctx := t.Context()
-	client := validatormock.NewMockValidatorClient(ctrl)
-
-	v := &validator{
-		km:              newMockKeymanager(t, randKeypair(t)),
-		validatorClient: client,
-		duties: &ethpb.ValidatorDutiesContainer{
-			PrevDependentRoot: bytesutil.PadTo([]byte{0x01, 0x02, 0x03}, fieldparams.RootLength),
-			CurrDependentRoot: bytesutil.PadTo([]byte{0x04, 0x05, 0x06}, fieldparams.RootLength),
-			CurrentEpochDuties: []*ethpb.ValidatorDuty{
-				{
-					AttesterSlot:    params.BeaconConfig().SlotsPerEpoch,
-					ValidatorIndex:  200,
-					CommitteeIndex:  100,
-					CommitteeLength: 4,
-					PublicKey:       []byte("testPubKey_1"),
-					ProposerSlots:   []primitives.Slot{params.BeaconConfig().SlotsPerEpoch + 1},
-				},
-			},
-		},
-	}
-
-	t.Run("nil head event", func(t *testing.T) {
-		err := v.checkDependentRoots(ctx, nil)
-		require.ErrorContains(t, "received empty head event", err)
-	})
-
-	t.Run("invalid previous duty dependent root", func(t *testing.T) {
-		head := &structs.HeadEvent{
-			Slot:                      "0",
-			PreviousDutyDependentRoot: "invalid_hex",
-		}
-		err := v.checkDependentRoots(ctx, head)
-		require.ErrorContains(t, "failed to decode previous duty dependent root", err)
-	})
-
-	t.Run("invalid current duty dependent root", func(t *testing.T) {
-		head := &structs.HeadEvent{
-			Slot:                      "0",
-			PreviousDutyDependentRoot: "0x0102030000000000000000000000000000000000000000000000000000000000",
-			CurrentDutyDependentRoot:  "invalid_hex",
-		}
-		err := v.checkDependentRoots(ctx, head)
-		require.ErrorContains(t, "failed to decode current duty dependent root", err)
-	})
-
-	t.Run("update duties for previous root mismatch", func(t *testing.T) {
-		head := &structs.HeadEvent{
-			Slot:                      "1",
-			PreviousDutyDependentRoot: "0xe3f7a1b2c489d56f03a6b8d9c7e1fa2456bb09f3de42a67c8910fc3e7a5d4b12",
-			CurrentDutyDependentRoot:  "0xe3f7a1b2c489d56f03a6b8d9c7e1fa2456bb09f3de42a67c8910fc3e7a5d4b12",
-		}
-		client.EXPECT().SubscribeCommitteeSubnets(
-			gomock.Any(),
-			gomock.Any(),
-			gomock.Any(),
-		).DoAndReturn(func(_ context.Context, _ *ethpb.CommitteeSubnetsSubscribeRequest, _ []*ethpb.ValidatorDuty) (*emptypb.Empty, error) {
-			return nil, nil
-		}).AnyTimes()
-		client.EXPECT().Duties(gomock.Any(), gomock.Any()).Return(v.duties, nil)
-		err := v.checkDependentRoots(ctx, head)
-		require.NoError(t, err)
-	})
-
-	t.Run("update duties for current root mismatch", func(t *testing.T) {
-		head := &structs.HeadEvent{
-			Slot:                      "1",
-			PreviousDutyDependentRoot: "0x0102030000000000000000000000000000000000000000000000000000000000",
-			CurrentDutyDependentRoot:  "0xe3f7a1b2c489d56f03a6b8d9c7e1fa2456bb09f3de42a67c8910fc3e7a5d4b12",
-		}
-		client.EXPECT().Duties(gomock.Any(), gomock.Any()).Return(v.duties, nil)
-		var wg sync.WaitGroup
-		wg.Add(1)
-
-		client.EXPECT().SubscribeCommitteeSubnets(
-			gomock.Any(),
-			gomock.Any(),
-			gomock.Any(),
-		).DoAndReturn(func(_ context.Context, _ *ethpb.CommitteeSubnetsSubscribeRequest, _ []*ethpb.ValidatorDuty) (*emptypb.Empty, error) {
-			wg.Done()
-			return nil, nil
-		}).AnyTimes()
-		err := v.checkDependentRoots(ctx, head)
-		require.NoError(t, err)
-		util.WaitTimeout(&wg, 2*time.Second)
-	})
-	t.Run("no updates needed", func(t *testing.T) {
-		head := &structs.HeadEvent{
-			Slot:                      "0",
-			PreviousDutyDependentRoot: "0x0102030000000000000000000000000000000000000000000000000000000000",
-			CurrentDutyDependentRoot:  "0x0405060000000000000000000000000000000000000000000000000000000000",
-		}
-		curr, err := bytesutil.DecodeHexWithLength(head.CurrentDutyDependentRoot, fieldparams.RootLength)
-		require.NoError(t, err)
-		require.DeepEqual(t, curr, v.duties.CurrDependentRoot)
-		require.NoError(t, v.checkDependentRoots(ctx, head))
-	})
 }
 
 func TestGetAttestationData_PreElectraNoCaching(t *testing.T) {
